build: Add meson build
phosphor-ipmi-flash: Fix all warning for Meson build

TODO: Need to fix the issue with `alignof` on non-type in memcpyAligned.

Changes to note,
- `with_systemdsystemunitdir` and `with_tmpfilesdir` is removed since it
  is not being documented and used in OpenBMC.
- Removed Code coverage with `-DDHAVE_GCOV` does not use it and meson
  covers it.
- Removed `--enable-oe-sdk` for using the OpenBMC SDK. It should with
  directly with no change required.

+++ bmc/firmware-handler/test/firmware_commit_unittest.cpp
@@ -51,8 +51,11 @@ TEST_F(FirmwareHandlerCommitTest, VerifyCannotCommitOnFlashImage)
     EXPECT_CALL(*imageMock2, open("asdf", std::ios::out))
         .WillOnce(Return(true));
 
-    EXPECT_TRUE(handler->open(
-        0, blobs::OpenFlags::write | FirmwareFlags::UpdateFlags::ipmi, "asdf"));
+    EXPECT_TRUE(
+        handler->open(0,
+                      static_cast<std::uint16_t>(blobs::OpenFlags::write) |
+                          FirmwareFlags::UpdateFlags::ipmi,
+                      "asdf"));
 
     EXPECT_FALSE(handler->commit(0, {}));
 }
@@ -68,9 +71,11 @@ TEST_F(FirmwareHandlerCommitTest, VerifyCannotCommitOnHashFile)
     EXPECT_CALL(*imageMock1, open(StrEq(hashBlobId), std::ios::out))
         .WillOnce(Return(true));
 
-    EXPECT_TRUE(handler->open(
-        0, blobs::OpenFlags::write | FirmwareFlags::UpdateFlags::ipmi,
-        hashBlobId));
+    EXPECT_TRUE(
+        handler->open(0,
+                      static_cast<std::uint16_t>(blobs::OpenFlags::write) |
+                          FirmwareFlags::UpdateFlags::ipmi,
+                      hashBlobId));
 
     EXPECT_FALSE(handler->commit(0, {}));
 }
